chat messasing system done!

+++ RelationshipService.Application/Models/Match/Responses/MatchResponse.cs
@@ -4,7 +4,7 @@ public class MatchResponse
 {
     public string Name { get; set; } = string.Empty;
     public string PhotoUrl { get; set; } = string.Empty;
-    public int MatchId { get; set; }
+    public long MatchId { get; set; }
     public DateTime MatchDate { get; set; }
     public bool IsActive { get; set; }
 }
